pyproject.toml
[project]
name = "gemma-finetune-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "datasets>=2.19.0",
  "scikit-learn>=1.4.0",
  "transformers>=4.41.0",
  "peft>=0.11.0",
  "accelerate>=0.33.0",
  "torch>=2.7.1",
  "dspy>=3.0.3",
  "huggingface-hub[cli]>=0.34.4",
  "python-dotenv>=1.0.1",
  "ipykernel>=6.30.1",
  "notebook>=7.4.5",
  "rank-bm25>=0.2.2",
  "nltk>=3.9.1",
  "rapidfuzz>=3.13.0",
  "vllm>=0.6.3",
  "sentence-transformers>=5.1.0",
  "faiss-cpu>=1.12.0",
  "langchain>=0.3.27",
  "tavily-python",
  "langchain-community>=0.3.28",
  "langchain-huggingface>=0.3.1",
  "pypdf>=6.0.0",
  "wandb>=0.21.1",
  "mlflow>=3.8.1",
  "litellm>=1.75.7",
]
